Added Swagger Documentation Configuration
