+++ src/main/java/es/unizar/webeng/hello/SwaggerDoc.java
@@ -1,0 +1,22 @@
+package es.unizar.webeng.hello;
+
+import org.springframework.context.annotation.*;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+import org.springframework.context.annotation.Bean;
+
+@Configuration
+@EnableSwagger2
+public class SwaggerDoc {
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2)  
+            .select()
+            .apis(RequestHandlerSelectors.any())
+            .paths(PathSelectors.any())
+            .build();
+    }
+}
